requirements.txt
asgiref==3.5.2
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
install==1.3.5
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
typing-extensions==4.3.0
